requirements.txt
uvicorn==0.17.5
fastapi==0.75.0
Pillow==9.1.0
tensorflow==2.7.0